+++ app/Services/Transfers/AssetTransferService.php
@@ -5,10 +5,12 @@ namespace App\Services\Transfers;
 use App\DataTransferObjects\Transfers\AssetTransferData;
 use App\Enums\Workflows\LastAction;
 use App\Enums\Workflows\Status;
+use App\Facades\Elasticsearch;
 use App\Http\Requests\Transfers\AssetTransferRequest;
 use App\Models\Transfers\AssetTransfer;
 use App\Repositories\Transfers\AssetTransferRepository;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\DB;
 
 class AssetTransferService
@@ -23,9 +25,18 @@ class AssetTransferService
         return AssetTransfer::query()->where('nik', auth()->user()?->nik)->with('asset')->get();
     }
 
-    public function allToAssetTransferData()
+    public function allToAssetTransferData($search = null, $length = 50)
     {
-        return AssetTransferData::collection($this->all())->toCollection();
+        $data = Elasticsearch::setModel(AssetTransfer::class)
+            ->searchMultiMatch($search, $length)
+            ->all();
+
+        $assetTransferData = AssetTransferData::collection(
+            collect($data)->pluck('_source')
+        );
+
+        $userNik = auth()->user()?->nik;
+        return $assetTransferData->toCollection()->where('nik', $userNik);
     }
 
     public function updateOrCreate(AssetTransferRequest $request)
@@ -33,22 +44,31 @@ class AssetTransferService
         $data = AssetTransferData::from(array_merge($request->all(), ['status' => Status::OPEN]));
         DB::transaction(function () use ($data) {
             $assetTransfer = $this->assetTransferRepository->updateOrCreate($data);
+            if ($data->getKey()) {
+                $assetTransfer->workflows()->delete();
+            }
             TransferWorkflowService::setModel($assetTransfer)->store();
+            $this->assetTransferRepository->sendToElasticsearch($assetTransfer, $data->getKey());
         });
     }
 
     public function delete(AssetTransfer $assetTransfer)
     {
-        $assetTransfer->workflows()->delete();
-        return $assetTransfer->delete();
+        return DB::transaction(function () use ($assetTransfer) {
+            $assetTransfer->workflows()->delete();
+            $this->assetTransferRepository->deleteFromElasticsearch($assetTransfer);
+            return $assetTransfer->delete();
+        });
     }
 
     public static function getByCurrentApproval()
     {
-        $data = AssetTransfer::query()->with('asset')->whereHas('workflows', function (Builder $query) {
-            $query->where('last_action', LastAction::NOTTING)
-                ->where('nik', auth()->user()?->nik);
-        })->get();
+        $data = AssetTransfer::query()->with('asset')
+            ->whereHas('workflows', function (Builder $query) {
+                $query->where('last_action', LastAction::NOTTING)
+                    ->where('nik', auth()->user()?->nik);
+            })
+            ->get();
         return AssetTransferData::collection($data)->toCollection();
     }
 }
